requirements/base.txt
Django
psycopg2
djangorestframework
markdown
django-filter
numpy
django-rest-swagger
django-extensions